requirements.txt
streamlit
phidata
python-dotenv
groq
yfinance
duckduckgo-search
uvicorn